firebase setup
