+++ package.json
@@ -11,6 +11,7 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.2.3",
     "chart.js": "^4.2.0",
+    "firebase": "^9.16.0",
     "html-react-parser": "^3.0.8",
     "react": "^18.2.0",
     "react-chartjs-2": "^5.2.0",
