+++ src/CryptoContext.js
@@ -1,10 +1,24 @@
 import React, { createContext, useContext, useEffect, useState } from "react";
+import axios from 'axios';
+import { CoinList } from "./config/api";
 
 const Crypto = createContext();
 
 const CryptoContext = ({ children }) => {
   const [currency, setCurrency] = useState("INR");
   const [symbol, setSymbol] = useState("₹");
+  const [coins, setCoins] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [user, setUser] = useState(null);
+
+  const fetchCoins = async () => {
+    setLoading(true);
+    const { data } = await axios.get(CoinList(currency));
+    console.log(data);
+
+    setCoins(data);
+    setLoading(false);
+  };
 
   useEffect(() => {
     if (currency === "INR") setSymbol("₹");
@@ -12,7 +26,7 @@ const CryptoContext = ({ children }) => {
   }, [currency]);
 
   return (
-    <Crypto.Provider value={{ currency, setCurrency, symbol }}>
+    <Crypto.Provider value={{ currency, setCurrency, symbol , coins, loading,fetchCoins}}>
       {children}
     </Crypto.Provider>
   );
