+++ src/config/firebaseConfig.js
@@ -1,0 +1,10 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyBgaj0LGvuoKj0Qe-y2RGMPixhbsTnuq0g",
+    authDomain: "cryptotracker-56665.firebaseapp.com",
+    projectId: "cryptotracker-56665",
+    storageBucket: "cryptotracker-56665.appspot.com",
+    messagingSenderId: "565805976036",
+    appId: "1:565805976036:web:8030a0b6900fe06b761ff1"
+  };
+  
+  export default firebaseConfig;
