+++ src/firebase.js
@@ -1,0 +1,13 @@
+
+import {intializeApp} from 'firebase/app';
+import {getAuth} from 'firebase/auth';
+import {getFirestore} from 'firebase/firestore';
+import firebaseConfig from "./config/firebaseConfig";
+
+const firebaseApp = intializeApp(firebaseConfig);
+
+const auth = getAuth(firebaseApp);
+const db = getFirestore(firebaseApp);
+
+
+export {auth , db };
